pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sked-parser"
version = "1.1.0"
description = "Parses Ostfalia sked timetables into JSON"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "beautifulsoup4==4.12.3",
  "jsonpickle==3.3.0",
  "lxml==4.9.4",
  "pyyaml==6.0.2",
  "requests==2.32.3",
  "setuptools==75.1.0",
]
[project.license]
file = "LICENSE"

[project.scripts]
sked-parser = "sked_parser.__main__:main"

[tool.setuptools.packages.find]
include = ["sked_parser"]

[tool.ruff]
line-length = 140

[tool.uv]
dev-dependencies = ["pytest~=8.3.3", "requests-cache>=1.2.1", "ruff>=0.6.9"]
